feat: separa logica de middlewares e plugin da configuracao inicial do fasify
