+++ src/middlewares/index.ts
@@ -1,0 +1,6 @@
+import type { FastifyInstance } from "fastify";
+import { authenticationMiddleware } from "./auth.middleware";
+
+export function registerMiddlewares(app: FastifyInstance) {
+  authenticationMiddleware(app);
+}
