+++ src/plugins/cors.plugin.ts
@@ -1,0 +1,6 @@
+import fastifyCors from "@fastify/cors";
+import type { FastifyInstance } from "fastify";
+
+export function corsPlugin(app: FastifyInstance) {
+  app.register(fastifyCors, { origin: "*" });
+}
