+++ src/plugins/sweagger-ui.plugin.ts
@@ -1,0 +1,6 @@
+import fastifySwaggerUi from "@fastify/swagger-ui";
+import type { FastifyInstance } from "fastify";
+
+export function sweaggerUIFastifyPlugin(app: FastifyInstance) {
+  app.register(fastifySwaggerUi, { routePrefix: "/docs" });
+}
